app/requirements.txt
requests
atproto
python-dotenv
Pillow